+++ pypdfbox/filter/__init__.py
@@ -9,6 +9,7 @@ from .decode_result import DecodeResult
 from .filter import Filter
 from .filter_factory import FilterFactory
 from .flate_decode import FlateDecode
+from .flate_filter import FlateFilter
 from .jbig2_decode import JBIG2Decode
 from .jpx_decode import JPXDecode
 from .lzw_decode import LZWDecode
@@ -26,6 +27,7 @@ __all__ = [
     "Filter",
     "FilterFactory",
     "FlateDecode",
+    "FlateFilter",
     "JBIG2Decode",
     "JPXDecode",
     "LZWDecode",